Developing the Event System
I still had issues with spdlog not added when I clone it.
I'm adding a second repository same as this. The second will be a public repo. This private repo can't let the comminity help me with it so I needed a public repo.

Trello: https://trello.com/c/utGUWeMq/13-6-commits
Git stats